trustcontrol/requirements.txt
# ════════════════════════════════════════════════════════════
#  TrustControl — Зависимости сервера  (v3.0)
# ════════════════════════════════════════════════════════════

# Сервер
fastapi>=0.110.0
uvicorn[standard]>=0.27.0

# База данных
sqlalchemy>=2.0.25
aiosqlite>=0.19.0
asyncpg>=0.29.0
alembic>=1.13.0

# Авторизация
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt==4.0.1
python-multipart>=0.0.9

# OpenAI
openai>=1.30.0

# Anthropic
anthropic>=0.25.0

# Telegram (21+ с InlineKeyboard)
python-telegram-bot>=21.0

# Валидация данных
pydantic[email]>=2.7.0

# Config
python-dotenv>=1.0.0

# S3 / Supabase Storage
boto3>=1.34.0

